plugins/image-metadata-processor/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.5
    # via requests
idna==3.11
    # via requests
imageio==2.37.3
    # via scikit-image
jsonschema==4.26.0
    # via rembg
jsonschema-specifications==2025.9.1
    # via jsonschema
lazy-loader==0.5
    # via scikit-image
llvmlite==0.46.0
    # via numba
networkx==3.6.1
    # via scikit-image
numba==0.64.0
    # via pymatting
numpy==2.4.3
    # via
    #   imageio
    #   numba
    #   pymatting
    #   rembg
    #   scikit-image
    #   scipy
    #   tifffile
packaging==26.0
    # via
    #   lazy-loader
    #   pooch
    #   scikit-image
piexif==1.1.3
    # via -r requirements.in
pillow==12.1.1
    # via
    #   -r requirements.in
    #   imageio
    #   pymatting
    #   rembg
    #   scikit-image
platformdirs==4.9.4
    # via pooch
pooch==1.9.0
    # via rembg
pymatting==1.1.15
    # via rembg
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rembg==2.0.72
    # via -r requirements.in
requests==2.32.5
    # via pooch
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
scikit-image==0.26.0
    # via rembg
scipy==1.17.1
    # via
    #   pymatting
    #   rembg
    #   scikit-image
tifffile==2026.3.3
    # via scikit-image
tqdm==4.67.3
    # via rembg
urllib3==2.6.3
    # via requests
